// src/app/shared/interfaces.ts
export interface ICustomer{
    id: number;
    name: string;
    city: string;
    orderTotal?: number;
    customerSince: any;
}

export interface IOrder{
    customerId: number;
    orderItems: IOrderItem[];
}

export interface IOrderItem{
    id: number;
    productName: string;
    itemCost: number;

}